+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 Contributors (alphabetical): Matthias van der hallen, Pierre Carbonnelle.
 
 ### API
+* (breaking) Merge !82: Problem must be created with `extended=True` for `decision_table()` API
 
 ### Interactive Consultant
 * fix #89: `constructed from` types create error in IC